requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
APScheduler==3.10.4
telethon==1.36.0
mysql-connector-python==8.3.0
